requirements.txt
streamlit==1.49.1
agno==2.0.4
python-dotenv==1.1.1
googletrans==4.0.2
SpeechRecognition==3.14.3
pydub==0.25.1
Pillow==11.3.0
deep-translator==1.11.4
numpy
requests==2.32.5
pydantic==2.11.8
langchain-google-genai==2.1.10
openai==1.107.2
google-genai==1.36.0
pandas==2.3.2
langdetect-py==1.1.1
streamlit-drawable-canvas-fix==0.9.7
opencv-python
scikit-image
